# EICR_App/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CertMate (EICR-oMatic 3000) API
  version: 1.0.0
  description: |
    Backend API for CertMate — automated EICR/EIC electrical certificate creation
    for electrical inspectors. Powered by Express/Node.js with ES modules.

    ## Authentication
    Most endpoints require a Bearer JWT token obtained via `POST /api/auth/login`.
    Pass the token in the `Authorization` header: `Bearer <token>`.

    ## Certificate Types
    - **EICR** — Electrical Installation Condition Report (periodic inspection)
    - **EIC** — Electrical Installation Certificate (new installations)
  contact:
    name: CertMate Support
  license:
    name: Proprietary
servers:
  - url: https://certomatic3000.co.uk
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
...
security:
  - BearerAuth: []
paths:
...
  /api/job/{userId}/{jobId}/export/csv:
    get:
      tags: [Export]
      summary: Download circuits as CSV file
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "404":
          description: Job not found
  /api/job/{userId}/{jobId}/export/excel:
    get:
      tags: [Export]
      summary: Download full job data as Excel workbook
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Job not found
...
  /api/job/{userId}/{jobId}/email:
    post:
      tags: [Email & WhatsApp]
      summary: Send certificate via email
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to:
                  type: string
                  format: email
                clientName:
                  type: string
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "404":
          description: Job or PDF not found
        "503":
          description: Email not configured
...
  /api/job/{userId}/{jobId}/whatsapp:
    post:
      tags: [Email & WhatsApp]
      summary: Send certificate via WhatsApp
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phoneNumber]
              properties:
                phoneNumber:
                  type: string
      responses:
        "200":
          description: WhatsApp message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: Invalid phone number
        "503":
          description: WhatsApp not configured
...
  /api/calendar/auth-url:
    get:
      tags: [Calendar]
      summary: Get Google OAuth consent URL
      responses:
        "200":
          description: Auth URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
        "503":
          description: Calendar not configured
  /api/calendar/callback:
    post:
      tags: [Calendar]
      summary: Exchange Google OAuth code for tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Calendar connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Code required
        "503":
          description: Calendar not configured
...
  /api/calendar/events:
    get:
      tags: [Calendar]
      summary: Get upcoming inspection events
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      type: object
        "400":
          description: Calendar not connected
        "503":
          description: Calendar not configured
  /api/calendar/create-job-from-event:
    post:
      tags: [Calendar]
      summary: Create a job from a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [location]
              properties:
                summary:
                  type: string
                location:
                  type: string
                start:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  jobId:
                    type: string
                  address:
                    type: string
        "400":
          description: No location
...